requirements.txt
yfinance
pandas
scipy
plotly
dash
dash-bootstrap-components
python-dateutil